fix for typo in SimVars initBoolVar
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@25651 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/cpp/Core/System/SimVars.cpp
@@ -71,7 +71,7 @@ int& SimVars::initIntVar(size_t i)
 *  \param [in] i index in simvars memory
 *  \return simvar variable
 */
-bool& SimVars::initBoolVar(unsigned int i)
+bool& SimVars::initBoolVar(size_t i)
 {
   if (i < _dim_bool)
     return _bool_vars.get()->get()[i];
